refactor: use file system route api for posts
see: https://www.gatsbyjs.com/docs/reference/routing/file-system-route-api/
