+++ posts/2020-05-02-continuous-integration-with-netlify-build-plugins/2020-05-02-continuous-integration-with-netlify-build-plugins.mdx
@@ -1,7 +1,6 @@
 ---
 title: Continuous Integration With Netlify Build Plugins
 date: '2020-05-02T18:34:02.866Z'
-slug: 'continuous-integration-with-netlify-build-plugins'
 description: >
   With Netlify's new build plugins, we can hook into their build pipeline 
   and do fancy stuff like a simple continuous integration for every pull request.
