+++ posts/2020-05-17-collaborative-ui-development/2020-05-17-collaborative-ui-development.mdx
@@ -1,7 +1,6 @@
 ---
 title: Collaborative UI Development
 date: '2020-05-17T13:37:03.284Z'
-slug: 'collaborative-ui-development'
 ---
 
 !["illustration by some cool human"](./collaboration-illustration.png)
