+++ posts/2020-05-17-collaborative-ui-development/collaboration-illustration.png
